Use more `stdint` sized types, bounds-check loaded map data

+++ linuxdoom-1.10/r_main.c
@@ -805,7 +805,7 @@ R_PointInSubsector
 {
     node_t*	node;
     int		side;
-    int		nodenum;
+    unsigned short nodenum;
 
     // single subsector is a special case
     if (!numnodes)
